refactor(gastown): extract SCM operations and instrument AI calls
Moves PR status checking and PR feedback analysis from TownDO into a new town-scm
submodule. Instruments the Workers AI call for checking review threads with a new
api.external_request analytics event.
